implement reviewer comments
