+++ src/components/Home.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+const Home = () => (
+  <div className="home">
+    
+  </div>
+)
+
