+++ src/components/HotelCard.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+const HotelCard = () => (
+  <div className="hotel-card">
+    
+  </div>
+)
+
